==> 1_collecting_data.py <==
# -*- coding: utf-8 -*-
"""Desafio_ciclo2_1.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1wMd1PwH3yAbCXS_Eg6m0tR_OFi_-8IBP
"""

#--Coleta de Dados--
#imports
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.datasets import load_iris, make_multilabel_classification
from scipy import stats

from google.colab import files
uploaded = files.upload()

file = '/content/bootcamp_train.csv'
df = pd.read_csv(file)
df.head()

#coluna id --remoção
#(df.id == df.index).value_counts()
#df.columns
df = df.drop(columns=['id'])

#renomeio colunas
df = df.rename(columns={'id_produto':'id','temperatura_ar':'tempA','temperatura_processo':'tempP','umidade_relativa':'umiR','velocidade_rotacional':'velR','desgaste_da_ferramenta':'desgF','falha_maquina':'F','FDF (Falha Desgaste Ferramenta)':'FDF','FDC (Falha Dissipacao Calor)':'FDC','FP (Falha Potencia)':'FP','FTE (Falha Tensao Excessiva)':'FTE','FA (Falha Aleatoria)':'FAL'})
df.to_csv('bootcamp_train_1.csv')
